requirements.txt
Flask
dnspython
eventlet
flask_pymongo
flask_socketio
gridfs
pymongo
python-dotenv
requests
